+++ saic-java-mqtt-gateway/src/main/java/net/heberling/ismart/mqtt/carconfig/HVACSettings.java
@@ -1,0 +1,12 @@
+package net.heberling.ismart.mqtt.carconfig;
+
+public interface HVACSettings {
+
+  Integer getMinAllowedTemp();
+
+  Integer getMaxAllowedTemp();
+
+  int normalizeTemperature(Integer temp);
+
+  byte mapTempToSaicApi(Integer temp);
+}